requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.2.0
zenpy==2.0.46
